add and fix style on Section
