+++ client/src/App.css
@@ -9,6 +9,7 @@
 :root {
   --bgColor: #0d0d0d;
   --titleColor: #ed0000;
+  --bgRedColor: #7f1000;
   --textColorWhite: #fffef8;
   --titleFont: "Bagel Fat One";
   --textFont: "Open sans", sans-serif;
