+++ client/src/components/Section/Component-Section/PodcastItem.tsx
@@ -1,0 +1,20 @@
+import "./style.css";
+
+interface PodcastItemProps {
+  title: string;
+}
+
+export default function PodcastItem({ title }: PodcastItemProps) {
+  return (
+    <article className="component-item">
+      <a href="/podcasts">
+      <img
+        src="./src/assets/images/podcast.jpg"
+        alt={title}
+        className="component-image"
+      />
+      <figcaption>{title}</figcaption>
+      </a>
+    </article>
+  );
+};
