+++ client/src/components/Section/Component-Section/style.css
@@ -1,0 +1,24 @@
+.component-image {
+  border-radius: 8px;
+  width: 60%;
+}
+
+.component-item a:hover {
+    text-decoration: underline;
+}
+.component-item figcaption {
+  color: var(--textColorWhite);
+  font-family: var(--textFont);
+  font-size: medium;
+}
+
+@media screen and (max-width: 768px) {
+  .component-item {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  .component-image {
+    width: 100%;
+  }
+}
