+++ client/src/components/Section/SectionItems.css
@@ -1,21 +1,15 @@
-:root {
-  --bgColor: #0d0d0d;
-  --titleColor: #ed0000;
-  --textColorWhite: #fffef8;
-  --titleFont: "Bagel Fat One";
-  --textFont: "Open sans", sans-serif;
-}
-
 section.playlists-section {
   display: flex;
   justify-content: center;
-  padding: 2rem;
-  gap: 10rem;
-  background-color: #7f1000;
-  font-size: 24px;
+  padding: 3rem;
+  /* gap: 10rem; */
+  background-color: var(--bgRedColor);
+  font-size: 1.5rem;
   text-align: center;
   font-family: var(--textFont);
+  font-weight: bold;
 }
+
 
 @media screen and (max-width: 768px) {
   section.playlists-section {
@@ -23,6 +17,6 @@ section.playlists-section {
     flex-direction: column;
     justify-content: center;
     gap: 2rem;
-    background-color: #7f1000;
+    padding: 3rem;
   }
 }
