+++ client/src/pages/home/index.tsx
@@ -1,12 +1,14 @@
 import "../../App.css";
 import GetPlaylists from "../../components/displayBlocks/MyPlaylistSearch";
 import ImageBlockHome from "../../components/imageBlock/imageBlockHome";
+import SectionItems from "../../components/Section/SectionItems";
 
 export default function Home() {
   return (
     <main>
       <ImageBlockHome />
       <GetPlaylists />
+      <SectionItems />
     </main>
   );
 }
